refactor: move parser tests to src/parser_tests.rs
